requirements.txt
google-genai>=1.0.0
jinja2>=3.0.0
playwright>=1.40.0
pypdf>=4.0.0
python-dotenv>=1.0.0
